Renamed 'services' package into 'service'
