+++ patient-view-be/src/main/java/spvproject/patientviewbe/service/VitalSignService.java
@@ -1,4 +1,4 @@
-package spvproject.patientviewbe.services;
+package spvproject.patientviewbe.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
